tweak defaults. bugfix in run.

+++ src/run.py
@@ -300,8 +300,9 @@ if __name__ == "__main__":
                     step - 1, next(iter(trainer.val_loader)), "Validation"
                 )
                 print("Saving checkpoints...")
-                os.replace(os.path.join(out_dir, "ckpt.pt"),
-                           os.path.join(out_dir, "ckpt_last.pt"))
+                if os.path.exists(os.path.join(out_dir, "ckpt.pt")):
+                    os.replace(os.path.join(out_dir, "ckpt.pt"),
+                               os.path.join(out_dir, "ckpt_last.pt"))
                 torch.save(
                     {
                         "args": args,
